third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-kernelUnitLength-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'kernelUnitLength' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('kernelUnitLength') is "1"
PASS convolveMatrixElement.getAttribute('kernelUnitLength') is "0.05"
PASS successfullyParsed is true

TEST COMPLETE
